Move all LLMIntegration classes to internal/llmobs
