+++ ddtrace/contrib/langchain/patch.py
@@ -16,21 +16,18 @@ except ImportError:
 from pydantic import SecretStr
 
 from ddtrace import config
-from ddtrace.constants import ERROR_TYPE
-from ddtrace.contrib._trace_utils_llm import BaseLLMIntegration
 from ddtrace.contrib.langchain.constants import API_KEY
 from ddtrace.contrib.langchain.constants import COMPLETION_TOKENS
 from ddtrace.contrib.langchain.constants import MODEL
 from ddtrace.contrib.langchain.constants import PROMPT_TOKENS
-from ddtrace.contrib.langchain.constants import PROVIDER
 from ddtrace.contrib.langchain.constants import TOTAL_COST
-from ddtrace.contrib.langchain.constants import TYPE
 from ddtrace.contrib.langchain.constants import text_embedding_models
 from ddtrace.contrib.langchain.constants import vectorstore_classes
 from ddtrace.contrib.trace_utils import unwrap
 from ddtrace.contrib.trace_utils import with_traced_module
 from ddtrace.contrib.trace_utils import wrap
 from ddtrace.internal.agent import get_stats_url
+from ddtrace.internal.llmobs.llm_integrations import LangChainIntegration
 from ddtrace.internal.logger import get_logger
 from ddtrace.internal.utils import ArgumentError
 from ddtrace.internal.utils import get_argument_value
@@ -63,75 +60,6 @@ config._add(
         "_api_key": os.getenv("DD_API_KEY"),
     },
 )
-
-
-class _LangChainIntegration(BaseLLMIntegration):
-    _integration_name = "langchain"
-
-    def _set_base_span_tags(self, span, interface_type="", provider=None, model=None, api_key=None):
-        # type: (Span, str, Optional[str], Optional[str], Optional[str]) -> None
-        """Set base level tags that should be present on all LangChain spans (if they are not None)."""
-        span.set_tag_str(TYPE, interface_type)
-        if provider is not None:
-            span.set_tag_str(PROVIDER, provider)
-        if model is not None:
-            span.set_tag_str(MODEL, model)
-        if api_key is not None:
-            if len(api_key) >= 4:
-                span.set_tag_str(API_KEY, "...%s" % str(api_key[-4:]))
-            else:
-                span.set_tag_str(API_KEY, api_key)
-
-    @classmethod
-    def _logs_tags(cls, span):
-        # type: (Span) -> str
-        api_key = span.get_tag(API_KEY) or ""
-        tags = "env:%s,version:%s,%s:%s,%s:%s,%s:%s,%s:%s" % (  # noqa: E501
-            (config.env or ""),
-            (config.version or ""),
-            PROVIDER,
-            (span.get_tag(PROVIDER) or ""),
-            MODEL,
-            (span.get_tag(MODEL) or ""),
-            TYPE,
-            (span.get_tag(TYPE) or ""),
-            API_KEY,
-            api_key,
-        )
-        return tags
-
-    @classmethod
-    def _metrics_tags(cls, span):
-        # type: (Span) -> list
-        provider = span.get_tag(PROVIDER) or ""
-        api_key = span.get_tag(API_KEY) or ""
-        tags = [
-            "version:%s" % (config.version or ""),
-            "env:%s" % (config.env or ""),
-            "service:%s" % (span.service or ""),
-            "%s:%s" % (PROVIDER, provider),
-            "%s:%s" % (MODEL, span.get_tag(MODEL) or ""),
-            "%s:%s" % (TYPE, span.get_tag(TYPE) or ""),
-            "%s:%s" % (API_KEY, api_key),
-            "error:%d" % span.error,
-        ]
-        err_type = span.get_tag(ERROR_TYPE)
-        if err_type:
-            tags.append("%s:%s" % (ERROR_TYPE, err_type))
-        return tags
-
-    def record_usage(self, span, usage):
-        # type: (Span, Dict[str, Any]) -> None
-        if not usage or self.metrics_enabled is False:
-            return
-        for token_type in ("prompt", "completion", "total"):
-            num_tokens = usage.get("token_usage", {}).get(token_type + "_tokens")
-            if not num_tokens:
-                continue
-            self.metric(span, "dist", "tokens.%s" % token_type, num_tokens)
-        total_cost = span.get_metric(TOTAL_COST)
-        if total_cost:
-            self.metric(span, "incr", "tokens.total_cost", total_cost)
 
 
 def _extract_model_name(instance):
@@ -741,7 +669,7 @@ def patch():
     langchain._datadog_patch = True
 
     Pin().onto(langchain)
-    integration = _LangChainIntegration(
+    integration = LangChainIntegration(
         config=config.langchain,
         stats_url=get_stats_url(),
     )
