+++ ddtrace/internal/llmobs/llm_integrations/__init__.py
@@ -1,0 +1,5 @@
+from base import BaseLLMIntegration
+from langchain import LangChainIntegration
+from openai import OpenAIIntegration
+
+__all__ = ["BaseLLMIntegration", "LangChainIntegration", "OpenAIIntegration"]
