+++ ddtrace/internal/llmobs/llm_integrations/base.py
@@ -16,12 +16,13 @@ from ddtrace.internal.llmobs import LLMObsWriter
 from ddtrace.internal.log_writer import V2LogWriter
 from ddtrace.internal.utils.formats import asbool
 from ddtrace.sampler import RateSampler
+from ddtrace.settings import IntegrationConfig
 
 
 class BaseLLMIntegration:
     _integration_name = "baseLLM"
 
-    def __init__(self, config, stats_url):
+    def __init__(self, config: IntegrationConfig, stats_url: str) -> None:
         # FIXME: this currently does not consider if the tracer is configured to
         # use a different hostname. eg. tracer.configure(host="new-hostname")
         # Ideally the metrics client should live on the tracer or some other core
@@ -104,29 +105,28 @@ class BaseLLMIntegration:
         return self._span_pc_sampler.sample(span)
 
     def is_pc_sampled_log(self, span: Span) -> bool:
-        if not self._config.logs_enabled or not span.sampled:
+        if not self.logs_enabled or not span.sampled:
             return False
         return self._log_pc_sampler.sample(span)
 
     def is_pc_sampled_llmobs(self, span: Span) -> bool:
         # Sampling of llmobs payloads is independent of spans, but we're using a RateSampler for consistency.
-        if not self._config.llmobs_enabled:
+        if not self.llmobs_enabled:
             return False
         return self._llmobs_pc_sampler.sample(span)
 
     def start_log_writer(self) -> None:
-        if not self._config.logs_enabled:
+        if not self.logs_enabled:
             return
         self._log_writer.start()
 
     def start_llm_writer(self) -> None:
-        if not self._config.llmobs_enabled:
+        if not self.llmobs_enabled:
             return
         self._llmobs_writer.start()
 
     @abc.abstractmethod
-    def _set_base_span_tags(self, span, **kwargs):
-        # type: (Span, Dict[str, Any]) -> None
+    def _set_base_span_tags(self, span: Span, **kwargs) -> None:
         """Set default LLM span attributes when possible."""
         pass
 
@@ -146,13 +146,11 @@ class BaseLLMIntegration:
 
     @classmethod
     @abc.abstractmethod
-    def _logs_tags(cls, span):
-        # type: (Span) -> str
+    def _logs_tags(cls, span: Span) -> str:
         """Generate ddtags from the corresponding span."""
         pass
 
-    def log(self, span, level, msg, attrs):
-        # type: (Span, str, str, Dict[str, Any]) -> None
+    def log(self, span: Span, level: str, msg: str, attrs: Dict[str, Any]) -> None:
         if not self.logs_enabled:
             return
         tags = self._logs_tags(span)
@@ -166,20 +164,20 @@ class BaseLLMIntegration:
             "ddtags": tags,
         }
         if span is not None:
-            log["dd.trace_id"] = str(span.trace_id)
+            # FIXME: this is a temporary workaround until we figure out why 128 bit trace IDs are stored as decimals.
+            # log["dd.trace_id"] = str(span.trace_id)
+            log["dd.trace_id"] = "{:x}".format(span.trace_id)
             log["dd.span_id"] = str(span.span_id)
         log.update(attrs)
         self._log_writer.enqueue(log)
 
     @classmethod
     @abc.abstractmethod
-    def _metrics_tags(cls, span):
-        # type: (Span) -> list
+    def _metrics_tags(cls, span: Span) -> List[str]:
         """Generate a list of metrics tags from a given span."""
         return []
 
-    def metric(self, span, kind, name, val, tags=None):
-        # type: (Span, str, str, Any, Optional[List[str]]) -> None
+    def metric(self, span: Span, kind: str, name: str, val: Any, tags: Optional[List[str]] = None) -> None:
         """Set a metric using the context from the given span."""
         if not self.metrics_enabled:
             return
@@ -195,8 +193,7 @@ class BaseLLMIntegration:
         else:
             raise ValueError("Unexpected metric type %r" % kind)
 
-    def trunc(self, text):
-        # type: (str) -> str
+    def trunc(self, text: str) -> str:
         """Truncate the given text.
 
         Use to avoid attaching too much data to spans.
@@ -208,14 +205,17 @@ class BaseLLMIntegration:
             text = text[: self._config.span_char_limit] + "..."
         return text
 
-    def llm_record(self, span, attrs):
-        # type: (Span, Dict[str, Any]) -> None
+    def llm_record(self, span: Span, attrs: Dict[str, Any], tags: Optional[List[str]]) -> None:
         """Create a LLM record to send to the LLM Obs intake."""
         if not self.llmobs_enabled:
             return
-        llm_record = {}
-        if span is not None:
-            llm_record["dd.trace_id"] = str(span.trace_id)
-            llm_record["dd.span_id"] = str(span.span_id)
-        llm_record.update(attrs)
-        self._llmobs_writer.enqueue(llm_record)
+        llmobs_tags = self._llmobs_tags(span)
+        if span is not None and span.sampled:
+            # FIXME: this is a temporary workaround until we figure out why 128 bit trace IDs are stored as decimals.
+            llmobs_tags.insert(0, "dd.trace_id:{:x}".format(span.trace_id))
+            # llmobs_tags.insert(0, "dd.trace_id:{}".format(span.trace_id))
+            llmobs_tags.insert(1, "dd.span_id:{}".format(span.span_id))
+        if tags:
+            llmobs_tags += tags
+        attrs["ddtags"] = llmobs_tags
+        self._llmobs_writer.enqueue(attrs)
